requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asyncpg==0.29.0
attrs==23.2.0
autoflake==2.3.0
bcrypt==4.0.1
billiard==4.2.0
black==24.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.109.0
fastapi-cache2==0.2.1
fastapi-versioning==0.10.0
flake8==7.0.0
flower==2.0.1
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
humanize==4.9.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
jose==1.0.0
kombu==5.3.5
Mako==1.3.0
MarkupSafe==2.1.3
mccabe==0.7.0
multidict==6.0.5
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.26.4
orjson==3.9.10
packaging==23.2
pandas==2.2.1
passlib==1.7.4
pathspec==0.12.1
pendulum==3.0.0
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
prometheus-fastapi-instrumentator==6.1.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
pyasn1==0.5.1
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.6.1
pydantic-extra-types==2.4.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
pyflakes==3.2.0
pyright==1.1.352
pytest==8.0.1
pytest-asyncio==0.23.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2024.1
PyYAML==6.0.1
redis==4.6.0
rsa==4.9
sentry-sdk==1.40.6
setuptools==69.1.1
six==1.16.0
sniffio==1.3.0
sqladmin==0.16.0
SQLAlchemy==2.0.25
starlette==0.35.1
time-machine==2.13.0
tornado==6.4
typing_extensions==4.9.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.25.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
WTForms==3.1.2
yarl==1.9.4
